Update Payment 버그수정, 포인트 전액결제 가능, 결제창에서 최종 결제금액 나옴.

+++ src/main/java/com/onebook/frontapi/dto/payment/PaymentRequestDto.java
@@ -7,7 +7,7 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class PaymentRequestDto {
-    private long usedPoint;
+    private int usedPoint;
     private String method;
     private int totalAmount;
     private String currency;
